skills/go-photo-studio/scripts/requirements.txt
# Python dependencies for local identity scoring scripts.
# Used by:
# - deterministic-identity-score.py
# - embedding-identity-score.py
Pillow>=10.0.0
numpy>=1.24.0
onnxruntime>=1.18.0
insightface>=0.7.3
